V0.14.4 (#1378)

+++ autosklearn/metalearning/optimizers/metalearn_optimizer/metalearner.py
@@ -111,7 +111,8 @@ class MetaLearningOptimizer(object):
                 except KeyError:
                     # TODO should I really except this?
                     self.logger.info("Could not find runs for instance %s" % task_id)
-                    runs[task_id] = pd.Series([], name=task_id)
+                    runs[task_id] = pd.Series([], name=task_id, dtype=float)
+
             runs = pd.DataFrame(runs)
 
             kND.fit(all_other_metafeatures, runs)
